Struct hardware_wallet::TransactionInfo

pub struct TransactionInfo {
    pub nonce: U256,
    pub gas_price: U256,
    pub gas_limit: U256,
    pub to: Option<Address>,
    pub value: U256,
    pub data: Vec<u8>,
    pub chain_id: Option<u64>,
}

This is the transaction info we need to supply to Trezor message. It's more or less a duplicate of ethcore::transaction::Transaction, but we can't import ethcore here as that would be a circular dependency.

Fields
nonce: U256

Nonce

gas_price: U256

Gas price

gas_limit: U256

Gas limit

to: Option<Address>

Receiver

value: U256

Value

data: Vec<u8>

Data

chain_id: Option<u64>

Chain ID